Moved Field/Expression base classes
To avoid circular dependency issues, `Field` is used as input to `MetadataType`
constructor, `Expression` is output of `Field` methods, but then Field needs
`Range` which is defined in `model`
